+++ R/any_table.R
@@ -58,6 +58,9 @@
 #' Styles can be created with [excel_output_style()].
 #' @param output The following output formats are available: excel and excel_nostyle.
 #' @param na.rm FALSE by default. If TRUE removes all NA values from the variables.
+#' @param print_miss FALSE by default. If TRUE outputs all possible categories of the
+#' grouping variables based on the provided formats, even if there are no observations
+#' for a combination.
 #' @param print TRUE by default. If TRUE prints the output, if FALSE doesn't print anything. Can be used
 #' if one only wants to catch the output data frame and workbook with meta information.
 #' @param monitor FALSE by default. If TRUE, outputs two charts to visualize the functions time consumption.
@@ -311,6 +314,7 @@ any_table <- function(data_frame,
                       style          = .qol_options[["excel_style"]],
                       output         = .qol_options[["output"]],
                       na.rm          = .qol_options[["na.rm"]],
+                      print_miss     = .qol_options[["print_miss"]],
                       print          = .qol_options[["print"]],
                       monitor        = .qol_options[["monitor"]]){
 
@@ -646,7 +650,8 @@ any_table <- function(data_frame,
                              nesting    = "all",
                              types      = combinations,
                              notes      = FALSE,
-                             na.rm      = na.rm)) |>
+                             na.rm      = na.rm,
+                             print_miss = print_miss)) |>
             rename_pattern("pct_group", paste0("pct_group_", pct_group[1])) |>
             collapse::fsubset(TYPE != "total")
     }
@@ -770,7 +775,8 @@ any_table <- function(data_frame,
                                nesting    = "all",
                                types      = combinations,
                                notes      = FALSE,
-                               na.rm      = na.rm)) |>
+                               na.rm      = na.rm,
+                               print_miss = print_miss)) |>
                 rename_pattern("pct_group", paste0("pct_group_", pct_group[group])) |>
                 drop_type_vars()