+++ R/summarise_plus.R
@@ -42,6 +42,9 @@
 #' @param merge_back Newly summarised variables can be merged back to the original
 #' data frame if TRUE. Only works if nested = "deepest and no formats are defined.
 #' @param na.rm FALSE by default. If TRUE removes all NA values from the class variables.
+#' @param print_miss FALSE by default. If TRUE outputs all possible categories of the
+#' grouping variables based on the provided formats, even if there are no observations
+#' for a combination.
 #' @param monitor FALSE by default. If TRUE, outputs two charts to visualize the
 #' functions time consumption.
 #' @param notes TRUE by default. Prints notifications about NA values produced by
@@ -160,6 +163,7 @@ summarise_plus <- function(data_frame,
                            nesting    = "deepest",
                            merge_back = FALSE,
                            na.rm      = .qol_options[["na.rm"]],
+                           print_miss = .qol_options[["print_miss"]],
                            monitor    = .qol_options[["monitor"]],
                            notes      = TRUE){
     # Measure the time
@@ -553,10 +557,14 @@ summarise_plus <- function(data_frame,
             result_df <- original_df |>
                 collapse::fungroup() |>
                 collapse::join(result_df,
-                               on = group_vars,
-                               how = "left",
+                               on      = group_vars,
+                               how     = "left",
                                verbose = FALSE) |>
                 dropp(".temp_weight")
+        }
+        # If formats are applied and all format categories schould be output
+        else if (print_miss && !is.null(formats)){
+            result_df <- result_df |> print_miss(formats, group_vars)
         }
 
         #---------------------------------------------------------------------#
@@ -889,6 +897,11 @@ summarise_plus <- function(data_frame,
                     group_df[["TYPE"]]    <- type
                     group_df[["TYPE_NR"]] <- as.integer(index)
                     group_df[["DEPTH"]]   <- as.integer(i)
+
+                    # If formats are applied and all format categories schould be output
+                    if (print_miss && !is.null(formats)){
+                        group_df <- group_df |> print_miss(formats, combination)
+                    }
 
                     # Add data frame to list to add them together at the end
                     all_results[[type]] <- group_df
@@ -1245,4 +1258,105 @@ reorder_summarised_columns <- function(data_frame, requested_stats){
 
     # Put the ordered columns at the end of the data frame
     data_frame |> data.table::setcolorder(ordered_cols, after = collapse::fncol(data_frame))
+}
+
+###############################################################################
+# Handle what happens with missing categories, if they appear in a format
+###############################################################################
+#' Output Missing Values If They Are Present In A Format
+#'
+#' @description
+#' Sometimes it can be handy to not only output all categories which are present
+#' in a data frame but also additionally all categories which could have been
+#' potentialy present. A format can provide more categories which are NA in the
+#' data frame, this functions outputs them as NA rows.
+#'
+#' @param data_frame The result data frame which contains all formatte variables.
+#' @param formats The provided formats.
+#' @param group_vars All grouping variables.
+#'
+#' @return
+#' Returns a data frame with added missing categories.
+#'
+#' @noRd
+print_miss <- function(data_frame,
+                       formats,
+                       group_vars){
+    format_list <- formats
+
+    # First check if given variables to format are part of the data frame. This only
+    # can be the case if "formats" is provided but the variables are missing in the
+    # "class" parameter. In this case abort, because there shouldn't be extra categories,
+    # which shouldn't be there in the first place.
+    format_names <- names(formats)
+    format_vars  <- group_vars[group_vars %in% format_names]
+
+    if (length(format_vars) == 0){
+        return(data_frame)
+    }
+
+    # Determine which variables are actually formatted in the data frame and which
+    # have to be added unformatted.
+    no_format_vars <- group_vars[!group_vars %in% format_names]
+
+    # Make sure only variables stay in the format list which are part of the group variables.
+    # There is a missmatch, when generating all combinations. Some combinations have lesser
+    # grouping variables than formats are provided.
+    format_select <- group_vars
+
+    if (length(no_format_vars) > 0){
+        format_select <- format_select[!group_vars %in% no_format_vars]
+    }
+
+    format_list <- format_list[format_select]
+
+    # Normally I would check if there are missing categories at all and if not abort.
+    # I think this has no benefit here and only costs time, since I am only using this
+    # function as intended. I leave this comment here, to know what is missing, if a
+    # case comes up for which this check would be neccessary.
+
+    # Every unformatted variable is put into a list as single data frame with only
+    # its unique values. Only if unformatted variables are present.
+    if (length(no_format_vars) > 0){
+        no_format_list <- lapply(data_frame |> keep(no_format_vars),
+                                 function(variable){
+                                     no_format_df        <- data.table::as.data.table(collapse::funique(variable))
+                                     names(no_format_df) <- "label"
+
+                                     no_format_df
+                                 })
+
+        format_list <- c(format_list, no_format_list)
+    }
+
+    # Create the cartesian product of all given formats to generate all possible
+    # categories.
+    format_df <- suppressMessages(expand_formats(format_list)) |>
+        data.table::setcolorder(group_vars) |>
+        data.table::setorderv(group_vars)
+
+    # Generate pseudo variable to join
+    format_df[[".pseudo_join"]] <- 1
+
+    # Get values from automatically generated variables to add them back for NA
+    # values after the join
+    type_value    <- data_frame[1, "TYPE"]
+    type_nr_value <- data_frame[1, "TYPE_NR"]
+    depth_value   <- data_frame[1, "DEPTH"]
+
+    # Perform left join on all format categories, so that each possible category
+    # appears in the final data frame. Additionally take over the values from the
+    data_frame <- suppressMessages(
+        collapse::join(format_df, data_frame,
+                       on       = group_vars,
+                       how      = "left",
+                       verbose  = FALSE) |>
+            dropp(".pseudo_join"))
+
+    # Re apply automatically generated variables to get rid of NA values
+    data_frame[["TYPE"]]    <- type_value
+    data_frame[["TYPE_NR"]] <- type_nr_value
+    data_frame[["DEPTH"]]   <- depth_value
+
+    data_frame
 }