+++ tests/testthat/test-options.R
@@ -4,6 +4,7 @@ default_stat_labels   <- get_stat_labels()
 default_print         <- get_print()
 default_monitor       <- get_monitor()
 default_na            <- get_na.rm()
+default_print_miss    <- get_print_miss()
 default_output        <- get_output()
 default_titles        <- get_titles()
 default_footnotes     <- get_footnotes()
@@ -122,6 +123,15 @@ test_that("Set global na.rm option", {
 })
 
 
+test_that("Set global print_miss option", {
+    set_print_miss(TRUE)
+
+    new_options <- get_print_miss()
+
+    expect_equal(new_options, TRUE)
+})
+
+
 test_that("Set global output option", {
     set_output("excel")
 
@@ -152,17 +162,19 @@ test_that("Set global footnotes", {
 test_that("Reset global options", {
     reset_qol_options()
 
-    new_print     <- get_print()
-    new_monitor   <- get_monitor()
-    new_na        <- get_na.rm()
-    new_output    <- get_output()
-    new_titles    <- get_titles()
-    new_footnotes <- get_footnotes()
+    new_print      <- get_print()
+    new_monitor    <- get_monitor()
+    new_na         <- get_na.rm()
+    new_print_miss <- get_print_miss()
+    new_output     <- get_output()
+    new_titles     <- get_titles()
+    new_footnotes  <- get_footnotes()
 
-    expect_true(default_print     == new_print)
-    expect_true(default_monitor   == new_monitor)
-    expect_true(default_na        == new_na)
-    expect_true(default_output    == new_output)
+    expect_true(default_print      == new_print)
+    expect_true(default_monitor    == new_monitor)
+    expect_true(default_na         == new_na)
+    expect_true(default_print_miss == new_print_miss)
+    expect_true(default_output     == new_output)
     expect_equal(new_titles, NULL)
     expect_equal(new_footnotes, NULL)
 })
@@ -209,10 +221,11 @@ test_that("Abort setting global style options on empty list", {
 
 
 test_that("Abort setting global options on empty list", {
-    expect_message(set_print(1),     " X ERROR: Print option can only be TRUE or FALSE. Global option remains unchanged.")
-    expect_message(set_monitor(1),   " X ERROR: Monitor option can only be TRUE or FALSE. Global option remains unchanged.")
-    expect_message(set_na.rm(1),     " X ERROR: NA removal option can only be TRUE or FALSE. Global option remains unchanged.")
-    expect_message(set_output(1),    " X ERROR: Output can only be 'console', 'text', 'excel' or 'excel_nostyle'. Global option remains unchanged.")
-    expect_message(set_titles(1),    " X ERROR: Titles must be provided as character. Global titles remain unchanged.")
-    expect_message(set_footnotes(1), " X ERROR: Footnotes must be provided as character. Global footnotes remain unchanged.")
+    expect_message(set_print(1),      " X ERROR: Print option can only be TRUE or FALSE. Global option remains unchanged.")
+    expect_message(set_monitor(1),    " X ERROR: Monitor option can only be TRUE or FALSE. Global option remains unchanged.")
+    expect_message(set_na.rm(1),      " X ERROR: NA removal option can only be TRUE or FALSE. Global option remains unchanged.")
+    expect_message(set_print_miss(1), " X ERROR: Print missing categories option can only be TRUE or FALSE. Global option remains unchanged.")
+    expect_message(set_output(1),     " X ERROR: Output can only be 'console', 'text', 'excel' or 'excel_nostyle'. Global option remains unchanged.")
+    expect_message(set_titles(1),     " X ERROR: Titles must be provided as character. Global titles remain unchanged.")
+    expect_message(set_footnotes(1),  " X ERROR: Footnotes must be provided as character. Global footnotes remain unchanged.")
 })